pyproject.toml
[tool.poetry]
name = "ace-api"
version = "0.1.0"
description = "API for AI Cost Estimation Project"
authors = ["axcisd <axcisd@rit.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pre-commit = "^3.5.0"
alembic = "^1.12.1"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
passlib = {extras = ["bcrypyt"], version = "^1.7.4"}
pydantic-settings = "^2.0.3"
pydantic-extra-types = "^2.1.0"
phonenumbers = "^8.13.25"
email-validator = "^2.1.0.post1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"

[tool.poe]
poetry_command = ""

[tool.poe.tasks]
i = [
  { cmd = "poetry install" },
  { cmd = "pre-commit install" }
]
start = "uvicorn main:app --reload"
format = "black ."
format_check = "black --check ."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
